Vamos

+++ database/migrations/2024_10_04_120301_create_archivos_table.php
@@ -8,16 +8,18 @@ class CreateArchivosTable extends Migration
 {
     public function up()
     {
-        Schema::create('archivos', function (Blueprint $table) {
-            $table->id();
-            $table->string('nombre');
-            $table->text('contenido')->nullable();
-            $table->foreignId('propietario')->constrained('users')->onDelete('cascade');
-            $table->foreignId('carpeta_id')->constrained('carpetas')->onDelete('cascade');
-            $table->foreignId('version_actual_id')->nullable()->constrained('archivo_version')->onDelete('set null');
-            $table->timestamps();
-        });
-        
+        // Verificar si la tabla ya existe antes de crearla
+        if (!Schema::hasTable('archivos')) {
+            Schema::create('archivos', function (Blueprint $table) {
+                $table->id();
+                $table->string('nombre');
+                $table->text('contenido')->nullable();
+                $table->foreignId('propietario')->constrained('users')->onDelete('cascade');
+                $table->foreignId('carpeta_id')->constrained('carpetas')->onDelete('cascade');
+                $table->foreignId('version_actual_id')->nullable()->constrained('archivo_version')->onDelete('set null');
+                $table->timestamps();
+            });
+        }
     }
 
     public function down()
